requirements.txt
google-generativeai
requests
python-dotenv
firebase-admin
python-telegram-bot 
schedule
